requirements.txt
playwright==1.49.1
openpyxl==3.1.5
xlsxwriter==3.2.0
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
schedule==1.2.2
python-dateutil==2.9.0
pytest==8.3.4
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
